Use custom bitset

+++ extlib/ECS/lib/include/Component.h
@@ -14,13 +14,13 @@ public:
     const std::string& GetName() const {return _name;}
     virtual bool Serialize(nlohmann::json &ioJson) const = 0;
     virtual bool Read(const nlohmann::json &inJSON) = 0;
-    virtual size_t GetType() const = 0;
+    virtual uint16_t GetType() const = 0;
     virtual void Destroy() = 0;
 protected:
 	id _IDEntity = std::numeric_limits<id>::max();
 
     std::string _name = "";
-    static size_t family_counter;
+    static uint16_t family_counter;
 };
 
 
@@ -30,8 +30,8 @@ public:
     Component() {
     }
 
-    static size_t GetID() {
-        static size_t i = family_counter++;
+    static uint16_t GetID() {
+        static uint16_t i = family_counter++;
 
         return i;
     }
@@ -43,9 +43,9 @@ public:
     virtual bool Read(const nlohmann::json &inJSON) = 0;
     virtual void Destroy() = 0;
 
-    virtual size_t GetType() const
+    virtual uint16_t GetType() const
     {
-        static size_t type = GetID();
+        static uint16_t type = GetID();
         return type;
     }
