+++ extlib/ECS/lib/src/Component.cpp
@@ -1,6 +1,6 @@
 #include "Component.h"
 
-size_t BaseComponent::family_counter = 0;
+uint16_t BaseComponent::family_counter = 0;
 
 BaseComponent::BaseComponent() {}
